C#: Restructure existing CFG tests for try/finally
