+++ csharp/ql/test/library-tests/controlflow/graph/cflow.cs
@@ -143,117 +143,6 @@ public class ControlFlow
 
     public string this[int i] { get { return i + ""; } set { } }
 
-    void TryFinally()
-    {
-        try
-        {
-            Console.WriteLine("Try1");
-        }
-        finally
-        {
-            Console.WriteLine("Finally");
-        }
-
-        try
-        {
-            Console.WriteLine("Try2");
-            return; // Go to the finally block
-        }
-        catch (System.IO.IOException ex) when (true)
-        {
-            throw;  // Go to the finally block
-        }
-        catch (System.ArgumentException ex)
-        {
-            try
-            {
-                if (true) throw;
-            }
-            finally
-            {
-                throw new Exception("Boo!");
-            }
-        }
-        catch (Exception)
-        {
-        }
-        catch
-        {
-            return; // Dead
-        }
-        finally
-        {
-            Console.WriteLine("Finally");
-        }
-
-        try
-        {
-            Console.WriteLine("Try3");
-            return; // Go to the finally block
-        }
-        catch (System.IO.IOException ex) when (true)
-        {
-            throw;  // Go to the finally block
-        }
-        catch (Exception e) when (e.Message != null)
-        {
-        }
-        finally
-        {
-            Console.WriteLine("Finally");
-        }
-
-        var i = 10;
-        while (i > 0)
-        {
-            try
-            {
-                if (i == 0)
-                    return;
-                if (i == 1)
-                    continue;
-                if (i == 2)
-                    break;
-            }
-            finally
-            {
-                try
-                {
-                    if (i == 3)
-                        throw new Exception();
-                }
-                finally
-                {
-                    i--;
-                }
-            }
-        }
-
-        try
-        {
-            if (Field.Length == 0)
-                return;
-            if (Field.Length == 1)
-                throw new OutOfMemoryException();
-        }
-        finally
-        {
-            if (!(Field.Length == 0))
-                Console.WriteLine(Field);
-            if (Field.Length > 0)
-                Console.WriteLine(1);
-        }
-
-        try
-        {
-            var temp = 0 / System.Math.E;
-        }
-        catch
-        {
-            ; // dead
-        }
-    }
-
     void For()
     {
         int x = 0;
@@ -395,23 +284,6 @@ public class ControlFlowSub : ControlFlow
     ControlFlowSub(string s) : this() { }
 
     ControlFlowSub(int i) : this(i.ToString()) { }
-}
-
-class AbnormalFinallyExit
-{
-    void M()
-    {
-        try
-        {
-            Console.WriteLine("Try");
-        }
-        finally
-        {
-            throw new ArgumentException("");
-            Console.WriteLine("Dead");
-        }
-        Console.WriteLine("Dead");
-    }
 }
 
 class DelegateCall
